+++ Changelog.md
@@ -12,10 +12,10 @@
 - Replace the function `tas` of the `Assignment` model with a `tas` "has_many" association for that model (#6764)
 - Ensure starter files are passed to autotester in sorted order (#6771)
 - Ensure the "Random Incomplete Submission" button takes into account criteria assignment for TAs (#6800)
-- Ensure bonus marks are not included in assignment "out of" in submissions table (#6836)
-- Ensure assignment "out of" in submissions table is rounded to two decimal places (#6836)
 
 ## [v2.3.4]
+- Ensure bonus marks are not included in assignment "out of" in submissions table (#6836)
+- Ensure assignment "out of" in submissions table is rounded to two decimal places (#6836)
 - Added POST API for Group Creation (#6834)
 
 ## [v2.3.3]